Fix: Handle larger numbers, up to 19 digits

+++ etherscan-test/Helpers/NumberHelper.cs
@@ -3,15 +3,15 @@ namespace etherscan_test.Helpers
 {
 	public static class NumberHelper
 	{
-		public static string DecimalToHex(int number)
+		public static string DecimalToHex(UInt64 number)
 		{
 			return $"0x{number.ToString("X")}";
 		}
 
-		public static int HexToDecimal(string number)
+		public static UInt64 HexToDecimal(string number)
 		{
 			// Ignore the '0x' portion of the number
-			return Convert.ToInt32(number.Substring(2), 16);
+			return Convert.ToUInt64(number.Substring(2), 16);
 		}
 	}
 }
